Renaming stylesheets directory
