+++ components/EditOverlaySidebar/EditOverlaySidebar.module.scss
@@ -1,4 +1,4 @@
-@import '../../stylesheets/variables';
+@import '../../styles/variables';
 
 .EditOverlaySidebar {
   width: 300px;
