+++ components/Layout/Layout.module.scss
@@ -1,4 +1,4 @@
-@import '../../stylesheets/variables';
+@import '../../styles/variables';
 
 .Layout {
   header {
